+++ src/bin_agnostic_baker/main_agnostic_baker.ml
@@ -13,10 +13,11 @@ let () =
          let* retcode =
            Lwt_exit.wrap_and_exit
            @@ Client_main_run.lwt_run
-                ~disable_logging:false
+                ~disable_logging:true
                 (module Agnostic_baker_config)
                 ~select_commands:(fun _ _ ->
-                  Lwt_result_syntax.return @@ Commands.baker_commands)
+                  Lwt_result_syntax.return
+                  @@ Commands.baker_commands (module Agnostic_baker_config))
                 ()
          in
          Format.pp_print_flush Format.err_formatter () ;