+++ src/lib_agnostic_baker/commands.mli
@@ -5,9 +5,10 @@
 (*                                                                           *)
 (*****************************************************************************)
 
-(** [baker_commands] is the list of CLI commands available for
+(** [baker_commands client_config] is the list of CLI commands available for
     controlling a baker process in the Tezos client context. *)
 val baker_commands :
+  (module Client_main_run.M) ->
   Tezos_client_base.Client_context.full Tezos_clic.command list
 
 (** [accuser_commands] behaves similarly to [baker_commands], but for the