+++ CHANGES.rst
@@ -121,6 +121,9 @@ Signer
   returns ``Request to prove possession is not allowed`` otherwise.
   (MR :gl:`!18137`)
 
+- Enables daily logs by default, located under
+  ``<base-dir>/logs/octez-signer/``. (MR :gl:`!18429`)
+
 Baker
 -----
 
@@ -194,12 +197,18 @@ Accuser
 - Can now denounce double consensus operations where one or both
   involved operations are aggregates. (MR :gl:`!18091`)
 
+- Enables daily logs by default, located under
+  ``<base-dir>/logs/octez-accuser-<proto>/``. (MR :gl:`!18429`)
+
 Agnostic Accuser
 ----------------
 
 - Add ``octez-accuser`` agnostic accuser binary. This behaves in a similar way
   to the agnostic baker binary, automatically switching the underlying accuser
   process at protocol migration. (MR :gl:`!17738`)
+
+- Enables daily logs by default, located under
+  ``<base-dir>/logs/octez-accuser/``. (MR :gl:`!18429`)
 
 Proxy Server
 ------------